requirements.txt
fastapi>=0.121.2
uvicorn>=0.38.0
streamlit>=1.51.0
streamlit-folium>=0.25.2
pandas>=2.2.3
scikit-learn>=1.3.2
sentence-transformers>=5.1.2
typer>=0.20.0
geopandas>=1.1.1
shapely>=2.1.2
python-dotenv>=1.2.1
matplotlib>=3.10.7
termcolor>= 3.2.0